requirements.txt
asgiref==3.8.1
Django==5.1.6
django-admin==2.0.2
django-excel-response2==3.0.6
django-six==1.0.5
excel-base==1.0.4
gunicorn==23.0.0
isoweek==1.3.3
packaging==24.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
screen==1.0.1
six==1.17.0
sqlparse==0.5.3
TimeConvert==3.0.13
tzlocal==5.3
xlwt==1.3.0
python-dotenv
whitenoise
requests 
beautifulsoup4
pillow